#231 Refactor: add type for feature login
