+++ typescript/big-practice/src/scripts/types/user.ts
@@ -1,0 +1,6 @@
+interface User {
+  email: string;
+  password: string;
+}
+
+export default User;
